backend/requirements.txt
# Dependências do backend (fase 2)
fastapi==0.100.0
uvicorn==0.22.0
python-dotenv==1.0.0
openai>=1.0.0
